Summersday Theme Notes
Theme is almost ready for release. The current issue is when you try to
load a <script> it will not load. This theme needs all bootstrap
functionality to be working which will include a closable sidebar. We
have tried some bootsnipp code but it has been unsuccessful as
somethings stopping code from loading. For example the new zoom search
from bootsnipp. More features need adding to the theme. Admin options
need to be added. We will have a credits section in the admin panel for
anyone who commits to this project. This will be  a FREE Theme released
for OSClass. Search page needs updating,Dashboard needs completing, New
search needs adding to the homepage. Anything you can think of to make
this theme better. It is currently installable through a zip file.
Hopefully we will get this released soon? :)